GH-15: endret struktur etter fått inn en ny tab i Transaksjoner
